Unified bool "return" type (set/get) to use "byte" (uint8_t)

+++ libraries/MySensors/core/MyMessage.cpp
@@ -111,7 +111,7 @@ char* MyMessage::getString(char *buffer) const {
 }
 
 bool MyMessage::getBool() const {
-	return getInt();
+	return getByte();
 }
 
 uint8_t MyMessage::getByte() const {
